refactor: change naming part 2
