+++ src/index.js
@@ -1,8 +1,8 @@
-import { welcome } from "./modules/operation/welcome.js";
-import { exitApp } from "./modules/operation/exit.js";
-import { printDirectoryPath } from "./modules/operation/pwd.js";
+import { welcome } from "./modules/operation/welcome.mjs";
+import { exitApp } from "./modules/operation/exit.mjs";
+import { printDirectoryPath } from "./modules/operation/pwd.mjs";
 import { cli_reader } from "./modules/utils/readline.mjs";
-import { handleCommand } from "./modules/handlers/common_handler.js";
+import { handleCommand } from "./modules/handlers/common_handler.mjs";
 import { homedir } from 'node:os';
 
 const rl = cli_reader;
