+++ src/modules/handlers/file_ops_handler.mjs
